Fix a libusb support signedness problem. (st)
git-svn-id: svn://mielke.cc/main/brltty@5533 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/usb_libusb.c
@@ -191,7 +191,7 @@ usbReadEndpoint (
   if ((endpoint = usbGetInputEndpoint(device, endpointNumber))) {
     const UsbEndpointDescriptor *descriptor = endpoint->descriptor;
     UsbEndpointTransfer transfer = USB_ENDPOINT_TRANSFER(descriptor);
-    int result = -1;
+    ssize_t result = -1;
 
     switch (transfer) {
       case UsbEndpointTransfer_Bulk:
